+++ dotnet/src/SmartStandards.Logging/Logging/SmartStandards/InsLogger.WoSourceContext.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using Logging.SmartStandards.Internal;
+using System;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 
@@ -15,37 +14,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogTrace(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogDebug(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogInformation(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogWarning(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogError(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogCritical(currentSourceContext, 0, 0, messageTemplate, args);
     }
 
@@ -55,37 +54,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogTrace(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogDebug(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogInformation(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogWarning(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogError(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogCritical(currentSourceContext, 0, eventKindEnumElement, args);
     }
 
@@ -95,37 +94,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogTrace(currentSourceContext, 0, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogDebug(currentSourceContext, 0, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogInformation(currentSourceContext, 0, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogWarning(currentSourceContext, 0, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogError(currentSourceContext, 0, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       LogCritical(currentSourceContext, 0, ex);
     }
 
@@ -135,37 +134,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(0, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(1, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(2, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(3, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(4, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(long sourceLineId, int eventKindId, string messageTemplate, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(5, currentSourceContext, sourceLineId, eventKindId, messageTemplate, args);
     }
 
@@ -175,37 +174,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(0, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(1, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(2, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(3, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(4, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(long sourceLineId, Enum eventKindEnumElement, params object[] args) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(5, currentSourceContext, sourceLineId, eventKindEnumElement, args);
     }
 
@@ -215,37 +214,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(0, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(1, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(2, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(3, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(4, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(long sourceLineId, int eventKindId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(5, currentSourceContext, sourceLineId, eventKindId, ex);
     }
 
@@ -255,37 +254,37 @@ namespace Logging.SmartStandards {
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogTrace(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(0, currentSourceContext, sourceLineId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogDebug(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(1, currentSourceContext, sourceLineId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogInformation(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(2, currentSourceContext, sourceLineId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogWarning(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(3, currentSourceContext, sourceLineId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogError(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(4, currentSourceContext, sourceLineId, ex);
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     public static void LogCritical(long sourceLineId, Exception ex) {
-      string currentSourceContext = DevLogger.GetSourceContextNameForAssembly(Assembly.GetCallingAssembly());
+      string currentSourceContext = SourceContextDiscoverer.InferSourceContext(Assembly.GetCallingAssembly());
       Log(5, currentSourceContext, sourceLineId, ex);
     }